+++ mosaic-runtime/src/commonMain/kotlin/com/jakewharton/mosaic/ui/AnsiLevel.kt
@@ -1,0 +1,8 @@
+package com.jakewharton.mosaic.ui
+
+internal enum class AnsiLevel {
+	NONE,
+	ANSI16,
+	ANSI256,
+	TRUECOLOR,
+}