+++ samples/rrtop/src/main/kotlin/example/RrtopApp.kt
@@ -1,10 +1,12 @@
 package example
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import com.jakewharton.mosaic.LocalTerminal
+import com.jakewharton.mosaic.layout.background
 import com.jakewharton.mosaic.layout.fillMaxWidth
 import com.jakewharton.mosaic.layout.height
 import com.jakewharton.mosaic.layout.padding
@@ -15,7 +17,6 @@ import com.jakewharton.mosaic.text.buildAnnotatedString
 import com.jakewharton.mosaic.text.withStyle
 import com.jakewharton.mosaic.ui.Alignment
 import com.jakewharton.mosaic.ui.Box
-import com.jakewharton.mosaic.ui.Color
 import com.jakewharton.mosaic.ui.Text
 import example.common.Key
 import example.common.key
@@ -30,44 +31,47 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 
 @Composable
-fun RrtopApp(rrtopViewModel: RrtopViewModel) {
+fun RrtopApp(rrtopViewModel: RrtopViewModel, colorsPalette: RrtopColorsPalette) {
 	val terminal = LocalTerminal.current
 	val rrtopUiState by rrtopViewModel.uiStateFlow.collectAsState()
-	Box(
-		modifier = Modifier
-			.width(terminal.size.width)
-			.height(terminal.size.height - 1), // subtraction of one is necessary, because there is a line with a cursor at the bottom, which moves up all the content
-	) {
-		when (rrtopUiState.currentScreen) {
-			RrtopUiState.Screen.Main -> MainScreen(
-				rrtopUiState.screenUiState as MainScreenUiState,
-				modifier = Modifier.padding(bottom = 1),
-			)
-			RrtopUiState.Screen.Cmd -> CmdScreen(
-				rrtopUiState.screenUiState as CmdScreenUiState,
-				modifier = Modifier.padding(bottom = 1),
-			)
-			RrtopUiState.Screen.Stat -> StatScreen(
-				rrtopUiState.screenUiState as StatScreenUiState,
-				modifier = Modifier.padding(bottom = 1),
-			)
-			RrtopUiState.Screen.Slow -> SlowScreen(
-				rrtopUiState.screenUiState as SlowScreenUiState,
-				modifier = Modifier.padding(bottom = 1),
-			)
-			RrtopUiState.Screen.Raw -> RawScreen(
-				rrtopUiState.screenUiState as RawScreenUiState,
-				modifier = Modifier.padding(bottom = 1),
+	CompositionLocalProvider(LocalRrtopColorsPalette provides colorsPalette) {
+		Box(
+			modifier = Modifier
+				.width(terminal.size.width)
+				.height(terminal.size.height - 1) // subtraction of one is necessary, because there is a line with a cursor at the bottom, which moves up all the content
+				.background(LocalRrtopColorsPalette.current.mainBg),
+		) {
+			when (rrtopUiState.currentScreen) {
+				RrtopUiState.Screen.Main -> MainScreen(
+					rrtopUiState.screenUiState as MainScreenUiState,
+					modifier = Modifier.padding(bottom = 1),
+				)
+				RrtopUiState.Screen.Cmd -> CmdScreen(
+					rrtopUiState.screenUiState as CmdScreenUiState,
+					modifier = Modifier.padding(bottom = 1),
+				)
+				RrtopUiState.Screen.Stat -> StatScreen(
+					rrtopUiState.screenUiState as StatScreenUiState,
+					modifier = Modifier.padding(bottom = 1),
+				)
+				RrtopUiState.Screen.Slow -> SlowScreen(
+					rrtopUiState.screenUiState as SlowScreenUiState,
+					modifier = Modifier.padding(bottom = 1),
+				)
+				RrtopUiState.Screen.Raw -> RawScreen(
+					rrtopUiState.screenUiState as RawScreenUiState,
+					modifier = Modifier.padding(bottom = 1),
+				)
+			}
+			BottomStatusBar(
+				currentScreen = rrtopUiState.currentScreen,
+				commonInfo = rrtopUiState.commonInfo,
+				modifier = Modifier.fillMaxWidth()
+					.height(1)
+					.padding(horizontal = 1)
+					.align(Alignment.BottomCenter),
 			)
 		}
-		BottomStatusBar(
-			currentScreen = rrtopUiState.currentScreen,
-			commonInfo = rrtopUiState.commonInfo,
-			modifier = Modifier.fillMaxWidth()
-				.height(1)
-				.padding(horizontal = 1)
-				.align(Alignment.BottomCenter),
-		)
 	}
 
 	LaunchedEffect(Unit) {
@@ -94,7 +98,7 @@ private fun BottomStatusBar(
 	commonInfo: String,
 	modifier: Modifier = Modifier,
 ) {
-	Box(modifier = modifier) {
+	Box(modifier = modifier.background(LocalRrtopColorsPalette.current.menuBg)) {
 		Text(
 			buildAnnotatedString {
 				RrtopUiState.Screen.entries.forEachIndexed { index, screen ->
@@ -105,19 +109,26 @@ private fun BottomStatusBar(
 						RrtopUiState.Screen.Slow -> "Slow"
 						RrtopUiState.Screen.Raw -> "Raw"
 					}
-					withStyle(SpanStyle(if (screen == currentScreen) Color.BrightWhite else Color.BrightBlack)) {
+					if (screen == currentScreen) {
+						withStyle(SpanStyle(LocalRrtopColorsPalette.current.menuHighlightFg)) {
+							append(screenName)
+						}
+					} else {
 						append(screenName)
 					}
 					if (index < RrtopUiState.Screen.entries.lastIndex) {
-						append(" | ")
+						withStyle(SpanStyle(LocalRrtopColorsPalette.current.menuDividerFg)) {
+							append(" | ")
+						}
 					}
 				}
 			},
 			modifier = Modifier.align(Alignment.CenterStart),
+			color = LocalRrtopColorsPalette.current.menuFg,
 		)
 		Text(
 			commonInfo,
-			color = Color.BrightBlack,
+			color = LocalRrtopColorsPalette.current.statusBarFg,
 			modifier = Modifier.align(Alignment.CenterEnd),
 		)
 	}