+++ samples/rrtop/src/main/kotlin/example/screens/MainScreen.kt
@@ -20,6 +20,7 @@ import com.jakewharton.mosaic.ui.Column
 import com.jakewharton.mosaic.ui.Row
 import com.jakewharton.mosaic.ui.Spacer
 import com.jakewharton.mosaic.ui.Text
+import example.LocalRrtopColorsPalette
 import example.MainScreenUiState
 import example.common.BorderedTitledBox
 import example.common.Table
@@ -46,9 +47,15 @@ fun MainScreen(uiState: MainScreenUiState, modifier: Modifier = Modifier) {
 
 @Composable
 private fun CpuSection(uiState: MainScreenUiState, modifier: Modifier = Modifier) {
-	BorderedTitledBox(title = "cpu", modifier = modifier) {
+	BorderedTitledBox(
+		title = "cpu",
+		titleColor = LocalRrtopColorsPalette.current.cpuTitleFg,
+		borderColor = LocalRrtopColorsPalette.current.cpuBorderFg,
+		modifier = modifier,
+	) {
 		Row {
 			val endOffsetForLabels = 7
+			val chartLineForeground = LocalRrtopColorsPalette.current.cpuChartLineFg
 			Column(
 				modifier = Modifier
 					.fillMaxWidth(0.6f)
@@ -58,14 +65,15 @@ private fun CpuSection(uiState: MainScreenUiState, modifier: Modifier = Modifier
 							row = height / 2,
 							column = 0,
 							string = "·".repeat(width - endOffsetForLabels),
-							foreground = Color.BrightBlack,
+							foreground = chartLineForeground,
 						)
 					},
 			) {
 				LabeledDottedPlot(
+					labelColor = LocalRrtopColorsPalette.current.cpuChartAxisFg,
 					plotInfo = PlotInfo(
 						items = uiState.items,
-						color = Color.Red,
+						color = LocalRrtopColorsPalette.current.cpuSysCpuDatasetFg,
 						numberFromItem = { it.sysCpuUtilization },
 						stringFromItem = { it.sysCpuUtilizationFormatted },
 					),
@@ -73,9 +81,10 @@ private fun CpuSection(uiState: MainScreenUiState, modifier: Modifier = Modifier
 					modifier = Modifier.fillMaxWidth().fillMaxHeight(0.5f),
 				)
 				LabeledDottedPlot(
+					labelColor = LocalRrtopColorsPalette.current.cpuChartAxisFg,
 					plotInfo = PlotInfo(
 						items = uiState.items,
-						color = Color.Green,
+						color = LocalRrtopColorsPalette.current.cpuUserCpuDatasetFg,
 						numberFromItem = { it.userCpuUtilization },
 						stringFromItem = { it.userCpuUtilizationFormatted },
 					),
@@ -96,13 +105,21 @@ private fun CpuSection(uiState: MainScreenUiState, modifier: Modifier = Modifier
 
 @Composable
 private fun CpuThroughputSubsection(uiState: MainScreenUiState, modifier: Modifier = Modifier) {
-	BorderedTitledBox(title = "throughput", modifier = modifier) {
+	BorderedTitledBox(
+		title = "throughput",
+		titleColor = LocalRrtopColorsPalette.current.throughputTitleFg,
+		borderColor = LocalRrtopColorsPalette.current.throughputBorderFg,
+		modifier = modifier,
+	) {
 		TitledSolidPlot(
 			firstTitle = "Total commands: ${uiState.items.lastOrNull()?.totalOperations ?: "0"}",
+			firstTitleColor = LocalRrtopColorsPalette.current.throughputTotalCommandsTextFg,
 			secondTitle = "          Op/s: ${uiState.items.lastOrNull()?.operationsPerSecondFormatted ?: "0"}",
+			secondTitleColor = LocalRrtopColorsPalette.current.throughputOpsTextFg,
 			plotInfo = PlotInfo(
 				items = uiState.items,
-				color = Color.Blue,
+				color = LocalRrtopColorsPalette.current.throughputSparklineFg,
+				baselineColor = LocalRrtopColorsPalette.current.throughputSparklineBaselineFg,
 				numberFromItem = { it.operationsPerSecond },
 				stringFromItem = { it.operationsPerSecondFormatted },
 			),
@@ -119,19 +136,23 @@ private fun CpuLegendInfo(
 ) {
 	Text(
 		buildAnnotatedString {
-			withStyle(SpanStyle(color = Color.Red)) {
+			withStyle(SpanStyle(color = LocalRrtopColorsPalette.current.cpuSysCpuText2Fg)) {
 				append("•")
 			}
-			append("sys cpu: ")
-			withStyle(SpanStyle(color = Color.Red)) {
+			withStyle(SpanStyle(color = LocalRrtopColorsPalette.current.cpuSysCpuText1Fg)) {
+				append("sys cpu: ")
+			}
+			withStyle(SpanStyle(color = LocalRrtopColorsPalette.current.cpuSysCpuText2Fg)) {
 				append(sysCpuUtilization)
 			}
 			append("   ")
-			withStyle(SpanStyle(color = Color.Green)) {
+			withStyle(SpanStyle(color = LocalRrtopColorsPalette.current.cpuUserCpuText2Fg)) {
 				append("•")
 			}
-			append("user cpu: ")
-			withStyle(SpanStyle(color = Color.Green)) {
+			withStyle(SpanStyle(color = LocalRrtopColorsPalette.current.cpuUserCpuText1Fg)) {
+				append("user cpu: ")
+			}
+			withStyle(SpanStyle(color = LocalRrtopColorsPalette.current.cpuUserCpuText2Fg)) {
 				append(userCpuUtilization)
 			}
 		},
@@ -141,14 +162,22 @@ private fun CpuLegendInfo(
 
 @Composable
 private fun NetworkSection(uiState: MainScreenUiState, modifier: Modifier = Modifier) {
-	BorderedTitledBox(title = "network", modifier = modifier) {
+	BorderedTitledBox(
+		title = "network",
+		titleColor = LocalRrtopColorsPalette.current.networkTitleFg,
+		borderColor = LocalRrtopColorsPalette.current.networkBorderFg,
+		modifier = modifier,
+	) {
 		Column {
 			TitledSolidPlot(
 				firstTitle = "Total rx: ${uiState.items.lastOrNull()?.totalRx ?: ""}",
+				firstTitleColor = LocalRrtopColorsPalette.current.networkRxTotalTextFg,
 				secondTitle = "    Rx/s: ${uiState.items.lastOrNull()?.rxPerSecondFormatted ?: ""}",
+				secondTitleColor = LocalRrtopColorsPalette.current.networkRxSTextFg,
 				plotInfo = PlotInfo(
 					items = uiState.items,
-					color = Color.Red,
+					color = LocalRrtopColorsPalette.current.networkRxSparklineFg,
+					baselineColor = LocalRrtopColorsPalette.current.networkRxSparklineBaselineFg,
 					numberFromItem = { it.rxPerSecond },
 					stringFromItem = { it.rxPerSecondFormatted },
 				),
@@ -156,10 +185,13 @@ private fun NetworkSection(uiState: MainScreenUiState, modifier: Modifier = Modi
 			)
 			TitledSolidPlot(
 				firstTitle = "Total tx: ${uiState.items.lastOrNull()?.totalTx ?: ""}",
+				firstTitleColor = LocalRrtopColorsPalette.current.networkTxTotalTextFg,
 				secondTitle = "    Tx/s: ${uiState.items.lastOrNull()?.txPerSecondFormatted ?: ""}",
+				secondTitleColor = LocalRrtopColorsPalette.current.networkTxSTextFg,
 				plotInfo = PlotInfo(
 					items = uiState.items,
-					color = Color.Blue,
+					color = LocalRrtopColorsPalette.current.networkTxSparklineFg,
+					baselineColor = LocalRrtopColorsPalette.current.networkTxSparklineBaselineFg,
 					numberFromItem = { it.txPerSecond },
 					stringFromItem = { it.txPerSecondFormatted },
 				),
@@ -171,14 +203,22 @@ private fun NetworkSection(uiState: MainScreenUiState, modifier: Modifier = Modi
 
 @Composable
 private fun MemorySection(uiState: MainScreenUiState, modifier: Modifier = Modifier) {
-	BorderedTitledBox(title = "memory", modifier = modifier) {
+	BorderedTitledBox(
+		title = "memory",
+		titleColor = LocalRrtopColorsPalette.current.memoryTitleFg,
+		borderColor = LocalRrtopColorsPalette.current.memoryBorderFg,
+		modifier = modifier,
+	) {
 		Column {
 			TitledSolidPlot(
 				firstTitle = " Max memory: ${uiState.items.lastOrNull()?.maxMemory ?: ""}",
+				firstTitleColor = LocalRrtopColorsPalette.current.memoryMaxMemoryTextFg,
 				secondTitle = "Used memory: ${uiState.items.lastOrNull()?.memoryUsageFormatted ?: ""}",
+				secondTitleColor = LocalRrtopColorsPalette.current.memoryUsedMemoryTextFg,
 				plotInfo = PlotInfo(
 					items = uiState.items,
-					color = Color.Blue,
+					color = LocalRrtopColorsPalette.current.memoryUsedMemorySparklineFg,
+					baselineColor = LocalRrtopColorsPalette.current.memoryUsedMemorySparklineBaselineFg,
 					numberFromItem = { it.memoryUsage },
 					stringFromItem = { it.memoryUsageFormatted },
 				),
@@ -186,10 +226,13 @@ private fun MemorySection(uiState: MainScreenUiState, modifier: Modifier = Modif
 			)
 			TitledSolidPlot(
 				firstTitle = " Frag ratio: ${uiState.items.lastOrNull()?.fragmentationRatio ?: ""}",
+				firstTitleColor = LocalRrtopColorsPalette.current.memoryFragRatioTextFg,
 				secondTitle = " RSS memory: ${uiState.items.lastOrNull()?.memoryRssFormatted ?: ""}",
+				secondTitleColor = LocalRrtopColorsPalette.current.memoryRssMemoryTextFg,
 				plotInfo = PlotInfo(
 					items = uiState.items,
-					color = Color.Green,
+					color = LocalRrtopColorsPalette.current.memoryRssMemorySparklineFg,
+					baselineColor = LocalRrtopColorsPalette.current.memoryRssMemorySparklineBaselineFg,
 					numberFromItem = { it.memoryRss },
 					stringFromItem = { it.memoryRssFormatted },
 				),
@@ -201,71 +244,82 @@ private fun MemorySection(uiState: MainScreenUiState, modifier: Modifier = Modif
 
 @Composable
 private fun StatSection(uiState: MainScreenUiState, modifier: Modifier = Modifier) {
-	BorderedTitledBox(title = "stat", modifier = modifier) {
+	BorderedTitledBox(
+		title = "stat",
+		titleColor = LocalRrtopColorsPalette.current.statTitleFg,
+		borderColor = LocalRrtopColorsPalette.current.statBorderFg,
+		modifier = modifier,
+	) {
 		Table(
 			TableData(uiState.items),
 			TableConfig(
-				listOf(
+				titleColor = LocalRrtopColorsPalette.current.statTableHeaderFg,
+				columnConfigs = listOf(
 					TableConfig.ColumnConfig.StringColumnConfig(
 						title = "time",
 						stringFromItem = { it.time },
+						valueColor = LocalRrtopColorsPalette.current.statTableRowTop1Fg,
 						weight = 2,
 					),
 					TableConfig.ColumnConfig.StringColumnConfig(
 						title = "op/s",
-						valueColor = Color.Blue,
 						stringFromItem = { it.operationsPerSecondFormatted },
+						valueColor = LocalRrtopColorsPalette.current.statTableRowTop2Fg,
 					),
 					TableConfig.ColumnConfig.StringColumnConfig(
 						title = "user",
-						valueColor = Color.Blue,
 						stringFromItem = { it.userCpuUtilizationFormatted },
+						valueColor = LocalRrtopColorsPalette.current.statTableRowTop2Fg,
 					),
 					TableConfig.ColumnConfig.StringColumnConfig(
 						title = "sys",
-						valueColor = Color.Blue,
 						stringFromItem = { it.sysCpuUtilizationFormatted },
+						valueColor = LocalRrtopColorsPalette.current.statTableRowTop2Fg,
 					),
 					TableConfig.ColumnConfig.StringColumnConfig(
 						title = "mem use",
 						stringFromItem = { it.memoryUsageFormatted },
+						valueColor = LocalRrtopColorsPalette.current.statTableRowTop1Fg,
 					),
 					TableConfig.ColumnConfig.StringColumnConfig(
 						title = "mem rss",
 						stringFromItem = { it.memoryRssFormatted },
+						valueColor = LocalRrtopColorsPalette.current.statTableRowTop1Fg,
 					),
 					TableConfig.ColumnConfig.StringColumnConfig(
 						title = "fr rati",
-						valueColor = Color.Blue,
 						stringFromItem = { it.fragmentationRatio },
+						valueColor = LocalRrtopColorsPalette.current.statTableRowTop2Fg,
 					),
 					TableConfig.ColumnConfig.StringColumnConfig(
 						title = "hit rate",
-						valueColor = Color.Blue,
 						stringFromItem = { it.hitRateFormatted },
+						valueColor = LocalRrtopColorsPalette.current.statTableRowTop2Fg,
 					),
 					TableConfig.ColumnConfig.ProgressColumnConfig(
-						filledColor = Color.Blue,
-						emptyColor = Color.BrightBlack,
+						filledColor = LocalRrtopColorsPalette.current.statTableRowGaugeFg,
+						emptyColor = LocalRrtopColorsPalette.current.statTableRowGaugeBg,
 						progressFromItem = { it.hitRate / 100.0f },
 					),
 					TableConfig.ColumnConfig.StringColumnConfig(
 						title = "keys",
 						stringFromItem = { it.totalKeys },
+						valueColor = LocalRrtopColorsPalette.current.statTableRowTop1Fg,
 					),
 					TableConfig.ColumnConfig.StringColumnConfig(
 						title = "exp",
 						stringFromItem = { it.keysToExpire },
+						valueColor = LocalRrtopColorsPalette.current.statTableRowTop1Fg,
 					),
 					TableConfig.ColumnConfig.StringColumnConfig(
 						title = "exp/s",
-						valueColor = Color.Blue,
 						stringFromItem = { it.keysToExpirePerSecond },
+						valueColor = LocalRrtopColorsPalette.current.statTableRowTop2Fg,
 					),
 					TableConfig.ColumnConfig.StringColumnConfig(
 						title = "evt/s",
-						valueColor = Color.Blue,
 						stringFromItem = { it.evictedKeysPerSecond },
+						valueColor = LocalRrtopColorsPalette.current.statTableRowTop2Fg,
 					),
 				),
 			),
@@ -277,12 +331,14 @@ private fun StatSection(uiState: MainScreenUiState, modifier: Modifier = Modifie
 private data class PlotInfo<T>(
 	val items: List<T>,
 	val color: Color,
+	val baselineColor: Color = color,
 	val numberFromItem: (T) -> Number,
 	val stringFromItem: (T) -> String,
 )
 
 @Composable
 private fun LabeledDottedPlot(
+	labelColor: Color,
 	plotInfo: PlotInfo<MainScreenUiState.MainItem>,
 	endOffsetForLabels: Int,
 	modifier: Modifier = Modifier,
@@ -303,11 +359,13 @@ private fun LabeledDottedPlot(
 						row = minValueRow,
 						column = width - minValueText.length,
 						string = minValueText,
+						foreground = labelColor,
 					)
 					drawText(
 						row = maxValueRow,
 						column = width - maxValueText.length,
 						string = maxValueText,
+						foreground = labelColor,
 					)
 				}
 			},
@@ -376,13 +434,15 @@ private fun DottedPlot(
 @Composable
 private fun TitledSolidPlot(
 	firstTitle: String,
+	firstTitleColor: Color,
 	secondTitle: String,
+	secondTitleColor: Color,
 	plotInfo: PlotInfo<MainScreenUiState.MainItem>,
 	modifier: Modifier = Modifier,
 ) {
 	Column(modifier = modifier) {
-		SolitPlotTitle(firstTitle, plotInfo.color)
-		SolitPlotTitle(secondTitle, plotInfo.color)
+		SolitPlotTitle(firstTitle, firstTitleColor)
+		SolitPlotTitle(secondTitle, secondTitleColor)
 		SolidPlot(plotInfo, Modifier.fillMaxSize())
 	}
 }
@@ -412,7 +472,7 @@ private fun SolidPlot(
 					row = height - 1,
 					column = 0,
 					string = "_".repeat(startColumn),
-					foreground = Color.BrightBlack,
+					foreground = plotInfo.baselineColor,
 				)
 				val minNullableValue = lastRange.minOrNull()
 				val maxNullableValue = lastRange.maxOrNull()